chore: refactor radar_object_tracker_node.cpp and related files
